(feat) added btnActions function
-> adds players once clicked on add players

-> restart starts a new game by eliminating the
board and elements and calling renderGame &
controller().newGame(), as well as playersOnClick
to add the event listeners again.

(fix) added cell as a variable that gets set
without value inside function display, and it
is change inside renderGame(), so the other functions
have access.

(fix) commented out the console initial logs as it
was getting repeated as we change the players name.

(fix) commented out printBoard as it is just redundant